+++ src/php/Run/Domain/Repl/StackFrame.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phel\Run\Domain\Repl;
+
+final readonly class StackFrame
+{
+    public function __construct(
+        public string $file,
+        public int $line,
+        public ?string $class,
+        public ?string $function,
+    ) {}
+}